chore: websocket 연결 시, 인증 처리 adapter 설정
